requirements.txt
asgiref==3.9.1
charset-normalizer==3.4.2
colorama==0.4.6
decorator==5.2.1
distro==1.9.0
Django==5.2.4
django-polymorphic==4.1.0
djangorestframework==3.16.0
dnspython==2.7.0
filelock==3.18.0
h11==0.16.0
idna==3.10
imageio==2.37.0
imageio-ffmpeg==0.6.0
jiter==0.10.0
moviepy==2.1.2
numpy==2.2.5
pillow==10.4.0
proglog==0.1.11
pydantic_core==2.33.2
PyMuPDF==1.25.3
python-dotenv==1.1.0
sniffio==1.3.1
sqlparse==0.5.3
tqdm==4.67.1
typing-inspection==0.4.1
typing_extensions==4.14.0
tzdata==2025.2
urllib3==2.4.0
